refactor: refactor MainLayout structure
