+++ frontend/src/app/layouts/MainLayout/MainLayout.styles.ts
