+++ frontend/src/app/layouts/MainLayout/index.ts
@@ -1,1 +1,1 @@
-export { MainLayout } from './ui'
+export { MainLayout } from './MainLayout'
