--- frontend/src/app/layouts/MainLayout/ui/index.ts
@@ -1,1 +1,0 @@
-export { MainLayout } from './MainLayout'
